Rebase AdaptiveLoopDetector.kt to match the updates
